Rename prayer conflict to prayer info
